Add max_random, max_uncertainty entry points
