+++ asreview/models/query/__init__.py
@@ -14,10 +14,11 @@
 
 from asreview.models.query.max import MaxQuery
 from asreview.models.query.mixed import MixedQuery
+from asreview.models.query.mixed import MaxRandomQuery
+from asreview.models.query.mixed import MaxUncertaintyQuery
 from asreview.models.query.uncertainty import UncertaintyQuery
 from asreview.models.query.random import RandomQuery
 from asreview.models.query.cluster import ClusterQuery
-from asreview.models.query.max_random import MaxRandomQuery
 from asreview.models.query.utils import get_query_model
 from asreview.models.query.utils import get_query_class
 from asreview.models.query.utils import list_query_strategies
