--- asreview/models/query/max_random.py
@@ -1,31 +1,0 @@
-# Copyright 2019-2021 The ASReview Authors. All Rights Reserved.
-#
-# Licensed under the Apache License, Version 2.0 (the "License");
-# you may not use this file except in compliance with the License.
-# You may obtain a copy of the License at
-#
-#     http://www.apache.org/licenses/LICENSE-2.0
-#
-# Unless required by applicable law or agreed to in writing, software
-# distributed under the License is distributed on an "AS IS" BASIS,
-# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-# See the License for the specific language governing permissions and
-# limitations under the License.
-
-from asreview.models.query.mixed import MixedQuery
-
-
-class MaxRandomQuery(MixedQuery):
-    """Mixed (95% Maximum and 5% Random) query strategy.
-
-    A mix of maximum and random query strategies with a mix ratio of 0.95.
-    At each query 95% of the instances would be sampled with the maximum
-    query strategy after which the remaining 5% would be sampled with
-    the random query strategy.
-    """
-
-    name = "max_random"
-    label = "Mixed (95% Maximum and 5% Random)"
-
-    def _query(self, X):
-        return MixedQuery
